1. 添加 sring_util.rs 与 unit_test.rs 两个文件; 2. 将字符串方法与测试代码进行分离,拆分到不同的文件及模块中;
